test: split out scope tests
