+++ actix-web-codegen/CHANGES.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
+- Add `#[scope]` macro.
 - Prevent inclusion of default `actix-router` features.
 - Minimum supported Rust version (MSRV) is now 1.72.
-- Add a scope macro that takes a path
 
 ## 4.2.2
